Move each media type to their own form components
