+++ templates/vue/src/components/node-modal/content-form/TextForm.vue
@@ -1,0 +1,21 @@
+<template>
+  <b-form-group label="Text content">
+    <b-form-textarea
+      id="node-text-content"
+      v-model="node.typeData.textContent"
+      data-testid="node-textContent"
+      placeholder="Enter text here"
+    ></b-form-textarea>
+  </b-form-group>
+</template>
+
+<script>
+export default {
+  props: {
+    node: {
+      type: Object,
+      required: true,
+    },
+  },
+}
+</script>
